+++ crates/matrix-sdk-indexeddb/src/state_store/mod.rs
@@ -20,6 +20,7 @@ use std::{
 use anyhow::anyhow;
 use async_trait::async_trait;
 use gloo_utils::format::JsValueSerdeExt;
+use growable_bloom_filter::GrowableBloom;
 use indexed_db_futures::prelude::*;
 use matrix_sdk_base::{
     deserialized_responses::RawAnySyncOrStrippedState,
@@ -388,6 +389,9 @@ impl IndexeddbStateStore {
             StateStoreDataKey::RecentlyVisitedRooms(user_id) => {
                 self.encode_key(keys::KV, (StateStoreDataKey::RECENTLY_VISITED_ROOMS, user_id))
             }
+            StateStoreDataKey::UtdHookManagerData => {
+                self.encode_key(keys::KV, StateStoreDataKey::UTD_HOOK_MANAGER_DATA)
+            }
         }
     }
 }
@@ -449,6 +453,10 @@ impl_state_store!({
                 .map(|f| self.deserialize_event::<Vec<String>>(&f))
                 .transpose()?
                 .map(StateStoreDataValue::RecentlyVisitedRooms),
+            StateStoreDataKey::UtdHookManagerData => value
+                .map(|f| self.deserialize_event::<GrowableBloom>(&f))
+                .transpose()?
+                .map(StateStoreDataValue::UtdHookManagerData),
         };
 
         Ok(value)
@@ -474,6 +482,9 @@ impl_state_store!({
                 &value
                     .into_recently_visited_rooms()
                     .expect("Session data not a recently visited room list"),
+            ),
+            StateStoreDataKey::UtdHookManagerData => self.serialize_event(
+                &value.into_utd_hook_manager_data().expect("Session data not UtdHookManagerData"),
             ),
         };