+++ crates/matrix-sdk-ui/CHANGELOG.md
@@ -9,7 +9,14 @@ Breaking changes:
 Bug fixes:
 
 - `UtdHookManager` no longer re-reports UTD events as late decryptions.
-  ([#3840](https://github.com/matrix-org/matrix-rust-sdk/pull/3840))
+  ([#3480](https://github.com/matrix-org/matrix-rust-sdk/pull/3480))
+
+Other changes:
+
+- `UtdHookManager` no longer reports UTD events that were already reported in a
+  previous session.
+  ([#3519](https://github.com/matrix-org/matrix-rust-sdk/pull/3519))
+
 
 # 0.7.0